requirements.txt
discord.py
bakapi-v2